Trait module_vault_registry_rpc_runtime_api::VaultRegistryApi

pub trait VaultRegistryApi<Block: BlockT, AccountId, Wrapped, Collateral, UnsignedFixedPoint>: Core<Block> where
    AccountId: Codec,
    Wrapped: Codec,
    Collateral: Codec,
    UnsignedFixedPoint: Codec, {
    fn get_vault_collateral(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        vault_id: AccountId
    ) -> Result<Result<BalanceWrapper<Collateral>, DispatchError>, ApiError> { ... }

    fn get_vault_collateral_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        vault_id: AccountId
    ) -> Result<Result<BalanceWrapper<Collateral>, DispatchError>, ApiError> { ... }

    fn get_vault_total_collateral(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        vault_id: AccountId
    ) -> Result<Result<BalanceWrapper<Collateral>, DispatchError>, ApiError> { ... }

    fn get_vault_total_collateral_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        vault_id: AccountId
    ) -> Result<Result<BalanceWrapper<Collateral>, DispatchError>, ApiError> { ... }

    fn get_total_collateralization(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<Result<UnsignedFixedPoint, DispatchError>, ApiError> { ... }

    fn get_total_collateralization_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<Result<UnsignedFixedPoint, DispatchError>, ApiError> { ... }

    fn get_first_vault_with_sufficient_collateral(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        amount: BalanceWrapper<Wrapped>
    ) -> Result<Result<AccountId, DispatchError>, ApiError> { ... }

    fn get_first_vault_with_sufficient_collateral_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        amount: BalanceWrapper<Wrapped>
    ) -> Result<Result<AccountId, DispatchError>, ApiError> { ... }

    fn get_first_vault_with_sufficient_tokens(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        amount: BalanceWrapper<Wrapped>
    ) -> Result<Result<AccountId, DispatchError>, ApiError> { ... }

    fn get_first_vault_with_sufficient_tokens_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        amount: BalanceWrapper<Wrapped>
    ) -> Result<Result<AccountId, DispatchError>, ApiError> { ... }

    fn get_premium_redeem_vaults(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<Result<Vec<(AccountId, BalanceWrapper<Wrapped>)>, DispatchError>, ApiError> { ... }

    fn get_premium_redeem_vaults_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<Result<Vec<(AccountId, BalanceWrapper<Wrapped>)>, DispatchError>, ApiError> { ... }

    fn get_vaults_with_issuable_tokens(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<Result<Vec<(AccountId, BalanceWrapper<Wrapped>)>, DispatchError>, ApiError> { ... }

    fn get_vaults_with_issuable_tokens_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<Result<Vec<(AccountId, BalanceWrapper<Wrapped>)>, DispatchError>, ApiError> { ... }

    fn get_vaults_with_redeemable_tokens(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<Result<Vec<(AccountId, BalanceWrapper<Wrapped>)>, DispatchError>, ApiError> { ... }

    fn get_vaults_with_redeemable_tokens_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<Result<Vec<(AccountId, BalanceWrapper<Wrapped>)>, DispatchError>, ApiError> { ... }

    fn get_issuable_tokens_from_vault(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        vault: AccountId
    ) -> Result<Result<BalanceWrapper<Wrapped>, DispatchError>, ApiError> { ... }

    fn get_issuable_tokens_from_vault_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        vault: AccountId
    ) -> Result<Result<BalanceWrapper<Wrapped>, DispatchError>, ApiError> { ... }

    fn get_collateralization_from_vault(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        vault: AccountId, 
        only_issued: bool
    ) -> Result<Result<UnsignedFixedPoint, DispatchError>, ApiError> { ... }

    fn get_collateralization_from_vault_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        vault: AccountId, 
        only_issued: bool
    ) -> Result<Result<UnsignedFixedPoint, DispatchError>, ApiError> { ... }

    fn get_collateralization_from_vault_and_collateral(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        vault: AccountId, 
        collateral: BalanceWrapper<Collateral>, 
        only_issued: bool
    ) -> Result<Result<UnsignedFixedPoint, DispatchError>, ApiError> { ... }

    fn get_collateralization_from_vault_and_collateral_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        vault: AccountId, 
        collateral: BalanceWrapper<Collateral>, 
        only_issued: bool
    ) -> Result<Result<UnsignedFixedPoint, DispatchError>, ApiError> { ... }

    fn get_required_collateral_for_wrapped(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        amount_btc: BalanceWrapper<Wrapped>
    ) -> Result<Result<BalanceWrapper<Collateral>, DispatchError>, ApiError> { ... }

    fn get_required_collateral_for_wrapped_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        amount_btc: BalanceWrapper<Wrapped>
    ) -> Result<Result<BalanceWrapper<Collateral>, DispatchError>, ApiError> { ... }

    fn get_required_collateral_for_vault(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        vault_id: AccountId
    ) -> Result<Result<BalanceWrapper<Collateral>, DispatchError>, ApiError> { ... }

    fn get_required_collateral_for_vault_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        vault_id: AccountId
    ) -> Result<Result<BalanceWrapper<Collateral>, DispatchError>, ApiError> { ... }
}

Provided methods

fn get_vault_collateral(
&self,
__runtime_api_at_param__: &BlockId<Block>,
vault_id: AccountId
) -> Result<Result<BalanceWrapper<Collateral>, DispatchError>, ApiError>

Get the vault’s collateral (excluding nomination)

fn get_vault_collateral_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
vault_id: AccountId
) -> Result<Result<BalanceWrapper<Collateral>, DispatchError>, ApiError>

Get the vault’s collateral (excluding nomination)

fn get_vault_total_collateral(
&self,
__runtime_api_at_param__: &BlockId<Block>,
vault_id: AccountId
) -> Result<Result<BalanceWrapper<Collateral>, DispatchError>, ApiError>

Get the vault’s collateral (including nomination)

fn get_vault_total_collateral_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
vault_id: AccountId
) -> Result<Result<BalanceWrapper<Collateral>, DispatchError>, ApiError>

Get the vault’s collateral (including nomination)

fn get_total_collateralization(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<Result<UnsignedFixedPoint, DispatchError>, ApiError>

Get the total collateralization of the system

fn get_total_collateralization_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<Result<UnsignedFixedPoint, DispatchError>, ApiError>

Get the total collateralization of the system

fn get_first_vault_with_sufficient_collateral(
&self,
__runtime_api_at_param__: &BlockId<Block>,
amount: BalanceWrapper<Wrapped>
) -> Result<Result<AccountId, DispatchError>, ApiError>

Get the first available vault with sufficient collateral to fulfil an issue request with the specified amount of Wrapped.

fn get_first_vault_with_sufficient_collateral_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
amount: BalanceWrapper<Wrapped>
) -> Result<Result<AccountId, DispatchError>, ApiError>

Get the first available vault with sufficient collateral to fulfil an issue request with the specified amount of Wrapped.

fn get_first_vault_with_sufficient_tokens(
&self,
__runtime_api_at_param__: &BlockId<Block>,
amount: BalanceWrapper<Wrapped>
) -> Result<Result<AccountId, DispatchError>, ApiError>

Get the first available vault with sufficient tokens to fulfil a redeem request

fn get_first_vault_with_sufficient_tokens_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
amount: BalanceWrapper<Wrapped>
) -> Result<Result<AccountId, DispatchError>, ApiError>

Get the first available vault with sufficient tokens to fulfil a redeem request

fn get_premium_redeem_vaults(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<Result<Vec<(AccountId, BalanceWrapper<Wrapped>)>, DispatchError>, ApiError>

Get all vaults below the premium redeem threshold, ordered in descending order of this amount

fn get_premium_redeem_vaults_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<Result<Vec<(AccountId, BalanceWrapper<Wrapped>)>, DispatchError>, ApiError>

Get all vaults below the premium redeem threshold, ordered in descending order of this amount

fn get_vaults_with_issuable_tokens(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<Result<Vec<(AccountId, BalanceWrapper<Wrapped>)>, DispatchError>, ApiError>

Get all vaults with non-zero issuable tokens, ordered in descending order of this amount

fn get_vaults_with_issuable_tokens_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<Result<Vec<(AccountId, BalanceWrapper<Wrapped>)>, DispatchError>, ApiError>

Get all vaults with non-zero issuable tokens, ordered in descending order of this amount

fn get_vaults_with_redeemable_tokens(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<Result<Vec<(AccountId, BalanceWrapper<Wrapped>)>, DispatchError>, ApiError>

Get all vaults with non-zero redeemable tokens, ordered in descending order of this amount

fn get_vaults_with_redeemable_tokens_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<Result<Vec<(AccountId, BalanceWrapper<Wrapped>)>, DispatchError>, ApiError>

Get all vaults with non-zero redeemable tokens, ordered in descending order of this amount

fn get_issuable_tokens_from_vault(
&self,
__runtime_api_at_param__: &BlockId<Block>,
vault: AccountId
) -> Result<Result<BalanceWrapper<Wrapped>, DispatchError>, ApiError>

Get the amount of tokens a vault can issue

fn get_issuable_tokens_from_vault_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
vault: AccountId
) -> Result<Result<BalanceWrapper<Wrapped>, DispatchError>, ApiError>

Get the amount of tokens a vault can issue

fn get_collateralization_from_vault(
&self,
__runtime_api_at_param__: &BlockId<Block>,
vault: AccountId,
only_issued: bool
) -> Result<Result<UnsignedFixedPoint, DispatchError>, ApiError>

Get the collateralization rate of a vault

fn get_collateralization_from_vault_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
vault: AccountId,
only_issued: bool
) -> Result<Result<UnsignedFixedPoint, DispatchError>, ApiError>

Get the collateralization rate of a vault

fn get_collateralization_from_vault_and_collateral(
&self,
__runtime_api_at_param__: &BlockId<Block>,
vault: AccountId,
collateral: BalanceWrapper<Collateral>,
only_issued: bool
) -> Result<Result<UnsignedFixedPoint, DispatchError>, ApiError>

Get the collateralization rate of a vault and collateral

fn get_collateralization_from_vault_and_collateral_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
vault: AccountId,
collateral: BalanceWrapper<Collateral>,
only_issued: bool
) -> Result<Result<UnsignedFixedPoint, DispatchError>, ApiError>

Get the collateralization rate of a vault and collateral

fn get_required_collateral_for_wrapped(
&self,
__runtime_api_at_param__: &BlockId<Block>,
amount_btc: BalanceWrapper<Wrapped>
) -> Result<Result<BalanceWrapper<Collateral>, DispatchError>, ApiError>

Get the minimum amount of collateral required for the given amount of btc with the current threshold and exchange rate

fn get_required_collateral_for_wrapped_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
amount_btc: BalanceWrapper<Wrapped>
) -> Result<Result<BalanceWrapper<Collateral>, DispatchError>, ApiError>

Get the minimum amount of collateral required for the given amount of btc with the current threshold and exchange rate

fn get_required_collateral_for_vault(
&self,
__runtime_api_at_param__: &BlockId<Block>,
vault_id: AccountId
) -> Result<Result<BalanceWrapper<Collateral>, DispatchError>, ApiError>

Get the amount of collateral required for the given vault to be at the current SecureCollateralThreshold with the current exchange rate

fn get_required_collateral_for_vault_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
vault_id: AccountId
) -> Result<Result<BalanceWrapper<Collateral>, DispatchError>, ApiError>

Get the amount of collateral required for the given vault to be at the current SecureCollateralThreshold with the current exchange rate

Loading content...